Total change of direction(ha)
I finally gave up on the maze idea.  The dc motors with no guidance are simply too unreliable to accurately retrace steps.  It was a good concept, but needs better equipment (steppers, perhaps more USDS)

The new function is to locate the IR light in a walled space and stop when it reaches it.  It does this pretty slowly, but pretty accurately.